requirements.txt
bleach==6.3.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4    
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
Flask==3.1.2
flask-cors==6.0.2
Flask-Limiter==4.1.1
flask-talisman==1.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.4.1
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pipreqs==0.4.13
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
Werkzeug==3.1.5
wrapt==2.0.1
yarg==0.1.10
gunicorn==21.2.0